# head: rework handling of non-seekable files