@RenjiSann 2025-03-13
> Changes give large performance improvement.

Can you attach a benchmark to show the improvement ?