@karlmcdowall 2025-03-14
> > Changes give large performance improvement.
> 
> Can you attach a benchmark to show the improvement ?

The performance improvements are seen when the input comes in via a pipe. Comparing current `head` with `head_improved` which contains my changes...

```
hyperfine "cat ./shakespeare.txt | ./target/release/head -c -100000" "cat ./shakespeare.txt | ./target/release/head_improved -c -100000"
Summary
  cat ./shakespeare.txt | ./target/release/head_improved -c -100000 ran
   17.93 ± 2.56 times faster than cat ./shakespeare.txt | ./target/release/head -c -100000


hyperfine "cat ./shakespeare.txt | ./target/release/head -n -100000" "cat ./shakespeare.txt | ./target/release/head_improved -n -100000"
Summary
  cat ./shakespeare.txt | ./target/release/head_improved -n -100000 ran
   24.38 ± 2.60 times faster than cat ./shakespeare.txt | ./target/release/head -n -100000
```


Also should note though that I want to add a bit more test code before anyone reviews these changes. Give me another day or so please :)