@karlmcdowall 2025-03-17
In addition to the main changes here, I also made a few changes to add/remove buffering (`BufReader`/`BufWriter`) through the `head.rs` file - I've remove some buffering that was redundant and added it where I think it's helpful. Here are the performance numbers before/after my changes on my machine (running Linux) to confirm that nothing has regressed...

| Command | Time Before (ms) | Time After (ms) | Improvement Ratio |
| --- | --- | --- | --- |
| `head -n 100000 ./shakespeare.txt` | 3.6 | 3.0 | **1.19** |
| `head -n -100000 ./shakespeare.txt` | 3.9 | 3.9 | 1 |
| `head -c 100000 ./shakespeare.txt` | 1.0 | 1.0 | 1 |
| `head -c -100000 ./shakespeare.txt` | 1.2 | 1.2 | 1 |
| `cat ./shakespeare.txt \| head -n 100000` | 4.0 | 3.5 | **1.15** |
| `cat ./shakespeare.txt \| head -n -100000` | 133.5 | 5.7 | **23** |
| `cat ./shakespeare.txt \| head -c 100000` | 1.4 | 1.4 | 1 |
| `cat ./shakespeare.txt \| head -c -100000` | 75.5 | 4.0 | **19** |

So, no regressions, some reasonable improvements (~15%) for the `... -n 100000` cases, and some really big improvements for the `cat ... -100000` cases.

If there are any other usecases people think I should run then please let me know.