@karlmcdowall 2025-03-18
> Appart from the typo. this looks good ^^ It's nice to see such performance improvement 🚀

Thanks for the review!